Add feed digesting algorithm
Given a set of stories, find digestable stories and return a set of
digested stories.

Currently only implementing the single-listing digest stories. The
current implementation strategy is:

1) iterate through the source feed and build a map tracking
   listing stories by listing id, action and actor id
2) iterate through the listing id/ story map tuples to create
   digest stories where appropriate

The way this is implemented it will only ever result in one feed
story of a particular type for a given listing id/user id. This seems
like a feature, but I could believe it is a bug.