Add mycroft, jmx support to help us understand what's happening in RT
Mycroft is a var inspector created by the relevance folks:

https://github.com/relevance/mycroft

It's a standalone server that lets us inspect the vars in the running
VM - we can use this to inspect the cache and other state inside RT.
Run it on port 4055 for now - fairly arbitrarily chosen.

Add basic JMX support: for now disable authentication and SSL on the
assumption that we'll ssh to the machine to connect to JMX. Run JMX on
port 4056 for now - also arbitrary.

Also, update lein, add benchmarking on the expiration thread and
tighten up the error logging code a bit to attempt to reclaim
stacktraces in prod.