Update feed building code to digest and filter stories before inserting them
Slight dissonance between how we choose the number of stories to
digest (6 hours usually, all 1000 stories in this case) which will
probably lead to some slightly weird feed behavior. It'd probably be
better to select a subset of the stories to digest based on the
digesting cache low score? Inclined to wait on that until we see how
weird it looks.