Fix bug with feed head replacement
When we add stories to a digest story in the cache, we update the
digest story's score to that of the latest story. If the digest story
happens to be the last story in the cache for that feed, this means
the lowest score in the cache is higher than the score of the digest
story that has already been written out. This means when we flush the
cache again we'll create a duplicate digest story. This sucks.

Fix by tracking the cache's min score directly.