Add migration hash to shard config server
Also add utilities for initiating a migration for a particular feed
and tests confirming feed state is correct on both the source and
destination mid-migration.

When a feed is added to the migrations hash, the sharding logic will
ensure writes happen to two redises rather than just one. These feeds
will be kept in sync for the duration of the time the feed is in the
migrations hash.