Set rt env to test more explicitly in tests
(test-background) installs an around wrapper that redefs
risingtide.core.env

Support setting additional background in this wrapper because
the around wrapper needs to go last to apply to the other
background functions. This may be problematic if we need
to add background functions lower down in the file, but this seems
to do the trick for now.