Add ability to run local cluster with thrift drpc server
This is a bit complicated. The default local drpc object is a purely
in-memory thing, but I need a real thrift server so we can use this
locally with brooklyn. Problem is, if we use the normal thrift drpc
server configuration some code in storm NPEs because it assumes one is
always using the local drpc object with a local cluster. To get around
this, implement a simple version of the thrift drpc service that just
calls in to the local drpc object.

With that, getting a local, fully operational feed building topology
is as simple as:

lein run -m risingtide.storm.core/run-local!

Also fix a few bugs with expiration that I noticed while getting this
working.

Ruby changes coming in a separate commit directly to staging - they're
backwards compatible, so let's get em in main.