Work around DigestFeed and *StorySets not being serializable
We want to pass them out of the feed builder, but they aren't
serializable. Naive attempts at implementing serialization convinced
me that it will take a more concerted/less frantic effort to get
right, so I'm working around it by passing a Seq out of the
BaseBatchBolt and adjusting the add-to-feed bolt to handle
seqs.

Also turn up the parallelism on the drpc feed builder. Those forces
serialization, which would have exposed this issue earlier, and is
probably a good idea anyway