Add object deduper, use it to reduce bolt memory footprints
Right now we store different instances of the same story in different
feeds. This is a bit silly, as stories are immutable. In C, we'd just
store a bunch of pointers to the same in-memory object.

Happily, we can duplicate this in Clojure by maintaining an
atom mapping values to instances. See the namespace documentation for
more information.