Add some exclusions to keep slf4j out of the non-storm build
I saw issues having two versions of slf4j around. Because storm is
provided by the cluster in production, I suspect this was happening
and preventing logging from happening.

That doesn't really make sense, because we only saw this fail when we
moved to console logging, but we're in crazy territory anyway tonight.