Add tag like scorer to the mix
We should be adding stories to feeds if the listing referred to by
the story is tagged with a tag the user has liked.

I also discovered that missing GROUP BY clauses are resulting in far fewer
non-zero scores than we'd expect, resulting in many stories not being added to
user feeds. This is unfortunate.