Big refactoring of interest bolts
These needed a bunch of love. Factor some non-storm helpers out to a
separate namespace and introduce a macro that makes "count scorer"
definitions much clearer and more succint. The bare bones of a scorer
are now much clearer - count scorers use a function that returns a
count of the number of relationships between one entity and a list of
entities and uses that function to return a score for each entity in
the list for the given story.